+++ src/utils/plugin-css/index.ts
@@ -4,5 +4,5 @@ export type { PluginScriptSource } from './path-finder';
 
 // Re-export functions
 export { findPluginScriptPath, findPluginScriptSource, isPluginBundled } from './path-finder';
-export { linkAndLoad, isCssLoaded, isCssImported } from './loader';
+export { linkAndLoad, isCssLoaded, isCssImported, whenCssImported } from './loader';
 export { pluginCSS } from './core';